+++ test/resources/language/terraform/validate/main.tf
@@ -1,0 +1,7 @@
+variable "region" {
+  type = string
+}
+
+locals {
+  region = "${var.region}"
+}